+++ build-virtualbox.ps1
@@ -51,7 +51,10 @@ Test-PluginInstalled "vagrant-winrm-syncedfolders"
 
 
 echo "Running Pester Tests"
-Invoke-Pester -OutputFile PesterTestResults.xml -OutputFormat NUnitXml -EnableExit
+$result = Invoke-Pester -OutputFile PesterTestResults.xml -OutputFormat NUnitXml -PassThru
+if ($result.FailedCount -gt 0) {
+  exit 1
+}
 
 echo "Running 'vagrant up --provider virtualbox'"
 vagrant up --provider virtualbox | Tee-Object -FilePath vagrant.log  #  --no-destroy-on-error --debug
